requirements.txt
langchain_community
langchain_openai
langchain_text_splitters
python-dotenv
qdrant_client
langchain
streamlit